Final review changes
- Pre commit changes

- Fixed labels and removed console logging for tap events

- Unified labelling styles for uvvis and xrd

- Remake lock file

- Move example data into sub dirs

- Add XRD docs stub

- Rename plotting modules with expectation they will merge eventually

- Docs tidy up

- Add README note about XRD and road map
