--- INSTALL.md
@@ -1,15 +1,0 @@
-# Installation
-
-We recommend you use [`uv`](https://astral.sh/uv) for managing virtual environments and Python versions.
-
-Once you have `uv` installed, you can clone this repository and install the package in a fresh virtual environment with:
-
-```
-git clone git@github.com:datalab-org/datalab-app-plugin-nsitu
-cd datalab-app-plugin-insitu
-uv sync --all-extras --dev
-```
-
-## Development installation
-
-You can activate `pre-commit` in your local repository with `uv run pre-commit install`.
