+++ docs/.pages
@@ -1,4 +1,3 @@
 nav:
 - Overview: index.md
-- Installation: install.md
 - reference
