--- docs/install.md
@@ -1,1 +1,0 @@
-../INSTALL.md
