+++ docs/reference/xrd.md
@@ -1,0 +1,3 @@
+# *In situ* X-ray Diffraction (XRD)
+
+::: datalab_app_plugin_insitu.apps.xrd
