+++ tests/test_inisitu_xrd_block.py
@@ -6,7 +6,9 @@ from datalab_api.utils import bokeh_from_json
 def test_block(pytestconfig):
     from datalab_app_plugin_insitu.apps.xrd.blocks import XRDInsituBlock
 
-    test_data_zip = Path(__file__).parent.parent / "example_data" / "XRD_example_insitu_data.zip"
+    test_data_zip = (
+        Path(__file__).parent.parent / "example_data" / "xrd" / "Example_in_situ_XRD_data.zip"
+    )
     block = XRDInsituBlock(item_id="test-xrd-insitu")
     block.data["xrd_folder_name"] = "43_up"
     block.data["time_series_folder_name"] = "log"
